+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-08 23:04:15
-// Commit: chore(perf): update dependencies and cleanup unused imports in perf
+// Updated at: 2026-08-10 16:44:29
+// Commit: feat(perf): add bi-directional event bus handler for perf
 
 class perfComponentState {
-  final String commitRef = '3489';
-  final String lastMessage = 'chore(perf): update dependencies and cleanup unused imports in perf';
-  final DateTime timestamp = DateTime.parse('2026-08-08T23:04:15+05:00');
+  final String commitRef = '3503';
+  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
+  final DateTime timestamp = DateTime.parse('2026-08-10T16:44:29+05:00');
 }